torchao support for in4/int8/float8 weight quant

+++ qasr/misc/misc_utils.py
@@ -11,7 +11,7 @@ def parse_args():
                          help='Max num of tokens to generate')
     parser.add_argument('--max_think_tokens', type=int, default=256,
                          help='Max num of tokens to generate')
-    parser.add_argument('--model_dtype', type=str, default='float32',
+    parser.add_argument('--model_dtype', type=str, default='float16',
                         choices=['auto', 'bfloat16', 'float16', 'float32'])
     parser.add_argument('--force_asr_language', type=str, default=None)
 
@@ -29,7 +29,7 @@ def parse_args():
     parser.add_argument('--dataset_path', type=str, default='hf-audio/esb-datasets-test-only-sorted')
     parser.add_argument('--dataset', type=str, default='voxpopuli')
     parser.add_argument('--split', type=str, default='test')
-    parser.add_argument('--data_dtype', type=str, default='float32',
+    parser.add_argument('--data_dtype', type=str, default='float16',
                          choices=['bfloat16', 'float16', 'float32'])
 
     parser.add_argument('--batch_size', type=int, default=64)
